requirements.txt
requests~=2.32.5
pillow~=12.0.0
selenium~=4.37.0
serverchan-sdk